requirements.txt
## The following requirements were added by pip freeze:
Django==3.2.19
pytz==2023.3
djangorestframework==3.14.0
urllib3==1.26.15
gunicorn==20.1.0
dj-database-url==2.0.0
whitenoise==5.0.1
